modify for my environment

+++ build_parcel.sh
@@ -12,18 +12,11 @@ set -ex
 PARCEL_DIR=LIVY-$1
 PARCEL=$PARCEL_DIR-$2.parcel
 
-# Build Livy
-[ ! -d ./livy ] && git clone https://github.com/cloudera/livy.git
-
-cd ./livy
-
-git checkout branch-0.3
-git checkout e36b6f5
-
-mvn -DskipTests -Dspark.version=2.0.0 -Dhadoop-version=2.6.0-cdh5.9.0 clean package
-
-# Prepare parcel
-cd ../
+if [ ! -d ./livy ]
+then
+    echo 'Please download and unzip Livy to `./livy/` directory'
+    exit 1
+fi
 
 [ ! -d ./$PARCEL_DIR ] && rm -rf ./$PARCEL_DIR
 
@@ -35,10 +28,10 @@ mkdir -p ./$PARCEL_DIR/rsc-jars
 cp -r ./livy/bin ./$PARCEL_DIR/
 cp -r ./livy/conf ./$PARCEL_DIR/
 cp ./log4j.properties ./$PARCEL_DIR/conf/log4j.properties
-cp ./livy/server/target/jars/*.jar ./$PARCEL_DIR/jars/
-cp ./livy/repl/scala-2.10/target/jars/*.jar ./$PARCEL_DIR/repl_2.10-jars/
-cp ./livy/repl/scala-2.11/target/jars/*.jar ./$PARCEL_DIR/repl_2.11-jars/
-cp ./livy/rsc/target/jars/*.jar ./$PARCEL_DIR/rsc-jars/
+cp ./livy/jars/*.jar ./$PARCEL_DIR/jars/
+cp ./livy/repl_2.10-jars/*.jar ./$PARCEL_DIR/repl_2.10-jars/
+cp ./livy/repl_2.11-jars/*.jar ./$PARCEL_DIR/repl_2.11-jars/
+cp ./livy/rsc-jars/*.jar ./$PARCEL_DIR/rsc-jars/
 
 cp -r parcel-src/meta $PARCEL_DIR/
